update entry filenames
