+++ webpack.common.js
@@ -3,21 +3,21 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
   entry: {
-    app: './src/index.js',
-    landing: './src/landing.js',
+    index: './src/index.js',
+    app: './src/app.js',
   },
   plugins: [
     new HtmlWebpackPlugin({
       title: 'Webpack 5 Boilerplate',
       template: path.resolve(__dirname, 'src', 'index.html'),
       filename: 'index.html',
-      chunks: ['app'],
+      chunks: ['index'],
     }),
     new HtmlWebpackPlugin({
       title: 'Webpack 5 Boilerplate',
-      template: path.resolve(__dirname, 'src', 'landing.html'),
-      filename: 'landing.html',
-      chunks: ['landing'],
+      template: path.resolve(__dirname, 'src', 'app.html'),
+      filename: 'app.html',
+      chunks: ['app'],
     }),
   ],
   output: {
